fix: surface unowned drain errors and split ownership tests
Propagate killUnowned Handle.Kill failures through the spawn lease so
BeginLoopStop joins them after pending bind/rebind wait channels close.
Split the oversized agent spawn ownership tests by stop/bind/rebind windows.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
